moved filters into helpers folder
