+++ app/src/views/backup/BackupInfo.vue
@@ -142,8 +142,8 @@
 
 <script>
 import api from '@/api'
-import { readableDate } from '@/filters/date'
-import { humanSize } from '@/filters/size'
+import { readableDate } from '@/helpers/filters/date'
+import { humanSize } from '@/helpers/filters/human'
 
 export default {
   name: 'BackupInfo',
