+++ app/src/views/diagnosis/Diagnosis.vue
@@ -108,7 +108,7 @@
 
 <script>
 import api from '@/api'
-import { distanceToNow } from '@/filters/date'
+import { distanceToNow } from '@/helpers/filters/date'
 
 export default {
   name: 'Diagnosis',
