+++ app/src/views/service/ServiceList.vue
@@ -34,7 +34,7 @@
 
 <script>
 import api from '@/api'
-import { distanceToNow } from '@/filters/date'
+import { distanceToNow } from '@/helpers/filters/date'
 
 export default {
   name: 'ServiceList',
